Adding the ability to observe whether search results are shown/hidden/received/cleared
This is not ideal and needs further work. There are many places where we are switching between search results and suggestion results and both of these are reffered to externally as a result view. This has lead to confusion about whether the observability of the search results should include the suggestion results too or just be for the actual search results themselves. I will need to clarify this before proceeding. I will also need to clean up some areas as it is becoming untidy and hard to follow. Also once we clear search results they still persist in the resultsTableDataSource which leads to some confusion about whether these results can be shown again or nuked from memory. Buddy: Sam